requirements.txt
Flask
Gunicorn
flask_sqlalchemy
python-dotenv